moved autolabel to handler
